# docs/swagger.yaml
basePath: /
definitions:
  CreateCommentRequest:
    properties:
      message:
        maxLength: 255
        minLength: 3
        type: string
      photo_id:
        minimum: 1
        type: integer
    required:
    - message
    - photo_id
    type: object
  CreateCommentResponse:
    properties:
      create_at:
        example: "2022-10-10T11:52:28.431369Z"
        type: string
      id:
        example: 1
        type: integer
      message:
        example: This is a comment
        type: string
      photo_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
  CreatePhotoResponse:
    properties:
      caption:
        example: my first photo
        type: string
      created_at:
        example: "2022-10-10T11:52:28.431369Z"
        type: string
      id:
        example: 1
        type: integer
      photo_url:
        example: https://images.unsplash.com
        type: string
      title:
        example: First Photo
        type: string
    type: object
...
  DeleteResponse:
    description: DeleteResponse Model
    properties:
      message:
        example: your account has been successfully deleted
        type: string
    type: object
...
  PhotoRequest:
    description: PhotoRequest Model
    properties:
      caption:
        example: my first photo
        type: string
      photo_url:
        example: https://images.unsplash.com
        type: string
      title:
        example: First Photo
        type: string
    required:
    - photo_url
    - title
    type: object
  PhotoResponse:
    description: PhotoResponse Model
    properties:
      caption:
        example: my first photo
        type: string
      created_at:
        example: 2017-07-22'T'16:28:55.444
        type: string
      id:
        example: 1
        type: integer
      photo_url:
        example: https://images.unsplash.com
        type: string
      title:
        example: First Photo
        type: string
      updated_at:
        example: 2017-07-22'T'16:28:55.444
        type: string
      user:
        $ref: '#/definitions/User'
      user_id:
        example: 1
        type: integer
    type: object
...
  UpdateCommentRequest:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  UpdateCommentResponse:
    properties:
      id:
        example: 1
        type: integer
      message:
        example: This is a comment
        type: string
      photo_id:
        example: 1
        type: integer
      updated_at:
        example: "2022-10-10T11:52:28.431369Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  UpdatePhotoResponse:
    properties:
      caption:
        example: my first photo
        type: string
      id:
        example: 1
        type: integer
      photo_url:
        example: https://images.unsplash.com
        type: string
      title:
        example: First Photo
        type: string
      updated_at:
        example: "2022-10-10T11:52:28.431369Z"
        type: string
    type: object
...
  User:
    description: User Model
    properties:
      email:
        example: admin@localhost
        type: string
      id:
        example: 1
        type: integer
      username:
        example: admin
        type: string
    type: object
  model.Comment:
    properties:
      created_at:
        example: 2017-07-22'T'16:28:55.444
        type: string
      id:
        example: 1
        type: integer
      message:
        example: This is a comment
        type: string
      photo:
        $ref: '#/definitions/PhotoResponse'
      photo_id:
        example: 1
        type: integer
      updated_at:
        example: 2017-07-22'T'16:28:55.444
        type: string
      user:
        $ref: '#/definitions/User'
      user_id:
        example: 1
        type: integer
    type: object
host: localhost:8000
info:
  contact:
    email: dev.abdianrizky@gmail.com
    name: Abdian Rizky
    url: https://linktr.ee/adonism2k
  description: Ini adalah dokumentasi API Contract untuk tugas Final Project Hactiv8.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Final Project API
  version: "1.0"
paths:
  /comments/:
    get:
      consumes:
      - application/json
      description: Get the current Comments data
      parameters:
      - description: Authorization
        in: header
        name: auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/model.Comment'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get All Comments
      tags:
      - Comment
    post:
      consumes:
      - application/json
      description: Create a Comment
      parameters:
      - description: Authorization
        in: header
        name: auth
        required: true
        type: string
      - description: Create Comment Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/CreateCommentResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Comment
      tags:
      - Comment
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Comment
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/DeleteResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Comment
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: Update the current Comment data
      parameters:
      - description: Authorization
        in: header
        name: auth
        required: true
        type: string
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Comment Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/UpdateCommentResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Comment
      tags:
      - Comment
  /photos/:
    get:
      consumes:
      - application/json
      description: Get All Photos
      parameters:
      - description: Authorization
        in: header
        name: auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/PhotoResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get All Photos
      tags:
      - Photo
    post:
      consumes:
      - application/json
      description: Create a Photo
      parameters:
      - description: Authorization
        in: header
        name: auth
        required: true
        type: string
      - description: Create Photo Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PhotoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/CreatePhotoResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Photo
      tags:
      - Photo
  /photos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a photo
      parameters:
      - description: Photo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/DeleteResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Photo
      tags:
      - Photo
    put:
      consumes:
      - application/json
      description: Update the current Photo data
      parameters:
      - description: Authorization
        in: header
        name: auth
        required: true
        type: string
      - description: Photo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Photo Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PhotoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/UpdatePhotoResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Photo
      tags:
      - Photo
...
securityDefinitions:
  "":
    description: Bearer Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"